add button to export face as texture
However, I want the texture to export in the manager, not on the user's end with a button to trigger this. `self` doesn't work how I want it to, so calling the export function in the manager class only generates a blank image. But calling it in a button in the ContentView will generate the texture as an image..